requirements.txt
# Release Toolkit Python Dependencies
pyyaml>=6.0.1
requests>=2.31.0
python-dotenv>=1.0.0
jinja2>=3.1.2
semver>=3.0.2
click>=8.1.7
tabulate>=0.9.0
pytest>=7.4.0
pytest-mock>=3.12.0
